consolidate html tests in one file
